backend/stream/tweet_producer/requirements.txt
requests
kafka-python
openai
pydantic